Buyer page implemented

+++ src/main/java/com/example/vmi/repository/BuyerRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.vmi.entity.Buyer;
 
-public interface BuyerRepository extends JpaRepository<Buyer, Long> {
+public interface BuyerRepository extends JpaRepository<Buyer, Integer> {
 
 }
